[trello-27]: Moved files to ds directory in genetic_distance app
