+++ app/genetic_distance/views.py
@@ -1,15 +1,17 @@
-from flask import render_template, request, jsonify, abort, Response
+from flask import Response
+from flask import abort
+from flask import jsonify
+from flask import render_template
+from flask import request
 
 from app.genetic_distance import genetic_distance
-from app.genetic_distance.utils.StandardDistance import StandardDistance
-from app.genetic_distance.utils.GeometricDistance import GeometricDistance
-from app.genetic_distance.utils.TakezakiNeiDistance import TakezakiNeiDistance
-
-from app.userpanel.models import Page
-
+from app.genetic_distance.ds.GeometricDistance import GeometricDistance
+from app.genetic_distance.ds.StandardDistance import StandardDistance
+from app.genetic_distance.ds.TakezakiNeiDistance import TakezakiNeiDistance
+from app.helpers.constants import GENETIC_DISTANCE
 from app.helpers.db_helper import add_calculation
 from app.helpers.db_helper import add_customer_activity
-from app.helpers.constants import GENETIC_DISTANCE
+from app.userpanel.models import Page
 
 
 @genetic_distance.route('/genetic-distance')
@@ -27,22 +29,23 @@ def get_data():
         distance_type = {
             'standard': StandardDistance(data),
             'geometric': GeometricDistance(data),
-            'takezaki-nei': TakezakiNeiDistance(data)
+            'takezaki-nei': TakezakiNeiDistance(data),
         }
 
         gen_distance = distance_type[distance_choice]
         gen_distance.calcuate_distances()
         gen_distance.build_matrix()
 
-        add_calculation(module_name=f'{GENETIC_DISTANCE}_{distance_choice}',
-                        user_data=data, result=gen_distance.get_matrix().tolist(), ip_address=request.remote_addr)
+        add_calculation(
+            module_name=f'{GENETIC_DISTANCE}_{distance_choice}',
+            user_data=data,
+            result=gen_distance.get_matrix().tolist(),
+            ip_address=request.remote_addr,
+        )
 
-        return jsonify({
-            'data': {
-                'dendro_base64': gen_distance.render_dendrogram(),
-                'matrix': gen_distance.redner_matrix()
-            }
-        })
+        return jsonify(
+            {'data': {'dendro_base64': gen_distance.render_dendrogram(), 'matrix': gen_distance.redner_matrix()}}
+        )
     except TypeError as e:
         abort(Response(f'Please check type of input data. {str(e)}', 409))
     except ValueError as e:
@@ -53,7 +56,4 @@ def get_data():
 
 @genetic_distance.context_processor
 def inject():
-    return {
-        'module_desc': Page.query.filter_by(slug=request.path).first()
-    }
-
+    return {'module_desc': Page.query.filter_by(slug=request.path).first()}
